Move reward calculation out of service backend.
Instead of calculating reward in the service backend, a new Reward
class is added to the python frontend and user to compute rewards
based on observation values generated by the service.

The goal of this change is twofold:

  (1) to simplify the backend service API by merging the
  GetObservation(), GetReward(), and TakeAction() operators into a
  single Step() call. This in turn enables more efficient usage as
  multiple calls can be replaced by a single Step() invocation.

  (2) to make it easier for users to define their own reward
  spaces. Like with observation spaces, learning techniques are
  sensitive to the reward space design decisions and so experimenting
  with reward shaping/scaling/design is a common use case. By
  implementing rewards in Python that simplifies the experimentation.

The new RPC Step() interface accepts a list of actions and a list of
observations. The service first applies the actions and then generates
the observations. The frontend python code then uses those
observations to compute a reward value.

One downside of this approach is that it means it is no longer to add
support for a new compiler without writing any Python, but this seems
like a reasonable tradeoff for the benefits.