Rename `cb` to `translate`.
The purpose this callback is to translate an observation from some
input type to an output type. E.g., from the Observation protocol
buffer generated by compiler service to a python type, or from a
python type to a different python type. Callbacks can be composed:

  proto = # ... from_service
  observation = derived.translate(base.translate(proto))

This renames `cb` to a slightly more descriptive name `translate`.