Bind new getter methods to the Observation view class.
When adding a space to the observation view, a getter method is bound
using the ID of the space. E.g.

    >>> env = gym.make("llvm-v0")
    >>> env.reset()
    >>> env.observation.Autophase()
    array([  0,   0,  46,  40,   3,  45,  16,   2,  14,  18,   1,   0,   0,
        67,   0,  63,   5,   4,  81,  76,  38,  51,  52,  45,  15,   0,
        14,  37,   0,   8,  59,   6,  63,  16,  35,  18,   0, 117,   0,
         0,   0,   4,  28,   0,   0,  61,   0,   0,   0,   0,  67, 414,
       266,  10,   0, 195])

This enables easier tab-completion of observation space names in a
python REPL, but since the methods are not known until runtime, an IDE
or static type checker will not know about them and so it will seem to
them to be a call to an unknown method.