Purge "eager" from codebase.
The terms "eager observation" and "eager reward" are more confusing
than informative. The 'default' assumption of Gym environments is that
they will provide and observation and a reward for every step(). The
fact that CompilerGym supports lazy observation/reward calculation is
an extension. Therefore 'default observation/reward' is a better name.

This patch removes "eager" from the vocabulary of CompilerGym. Where
necessary for clarity, "default" is used as a substitute.